Fixing skill hotkey level reset on update
Fixes https://github.com/MrAntares/Ragna.roBrowser/issues/6

Only updates should use ShortCut.setElement, so this will fix the issue. If somehow something else also uses this function, then it will make it impossible for it to change skill levels.

This also checks if the selected level is still valid. If the max level dropped below the selected, then resets it to the new max.