Fixed shader precision issues on some devices
Some GPUs use high precision as default float precision (eg: Nvidia, Adreno) but others use a lower precision or the defined one (eg: Mali).
This results in screen flickering on every Mediatek/Samsung SOC Phones.
To counter this we need to set the precision to highp. This might result in some performance loss, but only on low-end phones, while the filckering is a much worse problem on every affected device.