Disabled packet.call for now
It caused packet conflicts when multiple of the same type of packet followed each other. The following packets copied the previous one's values and caused a ton of problems like damage display values being the same for every hit monster under the same area damage.

Not sure if this will work in the long term, so for now only disabled it and not removed.
If nothing breaks will remove it later.