Shortcut guild skill handling
*UNOFFICIAL*

Officially the guild skills are not addable to the shortcut bar. 
But since this is stupid I decided to make it work. ¯\_(ツ)_/¯ 
This makes it possible to add guild skills to the shortcut bar.

Notable bug:
Skills only appear when the guild window is opened, because that is the first time the list of guild skills get requested. After that they stay. This can be fixed with a type 3 guild window request packet. Will fix later.