Implemented TarotCard effects & minor adjustments
The effect itself is a bit different from the official. Instead of zooming in and out then fading, it just zooms out then fades. It was easier this way. Might fix later, but more than enough for now. Looks good anyways.
Also, couldn't find the original sound anywhere in the GRFs, so it uses the amplify magical power skill sound, which is pretty similar. I guess the original is also this file with altered pitch or something other silly trick. If anyone finds the correct file name or method, please feel free to open an issue report.