Reworked EffectManager
When calling .spam instead of using separate arguments, you have to pass a (dynamic) param object with the call parameters. This object is passed on to the separate effect objects and duplicates and also to functions and renderers. This enables much better and cleaner parameter usage on every level of the rendering process with a ton more options that can be expanded/changed when needed without having to rework the argument passing everywhere it's used.