Fixed error when GL is not ready
Problem: Some packets like warlock summons add the effect before the map is ready, thus calling init on effect without GL being defined.
Solution: Only a needInit value is stored on effect add and the actual init is executed on the render function. The render function is only called when the renderer & gl is ready.